stdlib: make integer truncation a transparent function.
Integer truncation is a simple operation and should not need a ton of function calls including generation of metadata, even in Onone.
Making the corresponding initializer transparent dramatically improves performance of this operation with -Onone.
For optimized builds, there should be not much change, because originally, the initializer was inlinable + inline(always).

rdar://problem/64992353