[NFC] Write lnk file dynamically from cmake.
Instead of copying a static file from the repository, do what the
cmakefile already does for static-stdlib-args.lnk and generate it
dynamically. This means we can utilize the libicu flags set from the
prior commit rather than requiring libicu be built as part of the Swift
build just to ensure that the static link flags are correct.