[static][stdlib] Use the FindICU variables.
Instead of deriving the static library paths twice, use the variables generated by the FindICU cmake module.

The base cmakefiles only find icui18n and icuuc, but icudata is required to complete a static link, so we have to invoke find_package again here to ensure the variable is set.