Sema: Don't look through nested typealiases when checking for unsupported member reference
It appears that a long time ago, we didn't enforce that a member
reference to a typealias nested inside a generic type would supply
the generic arguments at all. To avoid breaking source compatibility,
we carved out some exceptions. Tighten up the exception to prohibit
the case where a typealias references another typealias, to fix a
crash.

While this worked in 5.1, it would crash in 5.2 and 5.3, and at
this point it's more trouble than it is worth to make it work
again, because of subtle representational issues. So let's just
ban it.

Fixes <rdar://problem/63535194>.