[Test] Specialize Prototypes/BigInt.swift for Windows.
Building the file with asserts crashes the compiler.  Building the file
with no_asserts does *not* crash the compiler.

For non-Windows platforms, it is sufficient to `XFAIL: asserts`.

That does not work on Windows thanks to Python < 3: the crash is not
handled by lit's `not` there as it is on other platforms.  To handle the
expected crash on Windows when building with asserts, not --crash is
required.  However on other platforms, not --crash breaks the test in
no_asserts builds.

Here, a version of the test's run script specialized for Windows is
broken out into a separate BigInt-windows.swift file.

rdar://problem/65251059