test: make test syntactically compatible with Python3
Python3 does not like the use of the unpacked tuple in the argument
list, change the function's arity and require the caller to expand the
tuple.

`map` now returns an iterator rather than a list, simplify the iteration
to unpack the elements in the list and use that to invoke the map'ed
method instead.