[OSLogOptimization][Tests] Make the OSLogMandatoryOptTest more liberal in admitting SIL changes.
This patch updates the OSLogMandatoryOptTest test suite so that
it only checks whether an expected format string appears in the SIL,
and omits checking that it is also passed to the _getGlobalStringTablePointer
builtin. The latter property is anyway necessary for the code to compile.
This is done to avoid hardcoding the begin_borrow and copy_value
instructions that may be introduced when the format string is passed
to the builtin. The begin_borrow instructions are omitted in release
builds while they appear in debug builds.

<rdar://problem/65251877>