[SIL] Add a "unique" flag to alloc_ref instructions.
Adds a new flag to alloc_ref instructions. This flag, "unique",
signifies that the reference is unique meaning it is not stored
elsewhere and may not be written to elsewhere. This means accesses of
this reference may be "static" instead of "dynamic".

Uniqueness of the reference is enforced by the SILVerifier. This is good
because it may be run between each pass so errors will be easy to track
down.

Currently the SILVerifier only supports a very few cases where it can
prove uniqueness, all other cases will end in an error. Later this
verifier check may be extended to be more robust.

Currently the only way to mark a reference as unique is through textual
SIL. Later optimization passes may be added or updated to add this
information.