Fix call for upstream llvm change of ClangDecl::isInLocalScope()
After f721e0582b158c60c56d2601235b6d60758f4d7a in llvm it's now called
isInLocalScopeForInstantiation() and returns false if the context is not
dependent on a template parameter, so doesn't have the desired behavior for the
single swift caller in the clang importer.

This pulls out the old logic into a new static isLocalScope() function in
ClangAdapter.cpp.