validation: make test_shell Python3 friendly
The custom `shell` module would coerce the byte-string representation
from `subprocess` into a text string representation in the happy path.
However, the same handling was not applied to the error case.  This
would result in a byte-string representation being returned in Python 3
rather than the text string.  Perform the conversion in both cases to
ensure that we can handle the strings without having to do the
conversion.