gyb: explicitly convert stream to UTF-8
@tbkka made the observation that the CI builders may be failing as
Python will default to using the environment's setting to decode the
file (which we expect to be in UTF-8).  If the environment does not have
`LC_ALL` set, it will use the fallback locale of `C` which is going to
use ASCII.  This results in the decoding failure.  Convert the argument
from `FileType` to `str` and use `io.open` to allow us to open the file
with the correct encoding and line-ending irrespective of the Python
version.