gyb: force I/O in UTF-8
We would rely on the system encoding, which if unspecified would fall
back to `C` treating the encoding as ASCII.  This fails with unicode
data in Python3.  Thanks to @tbkka for the brilliant idea of forcing the
encoding when reading the file.  This improves the test pass rate on
Linux with Python 3.