GSB: An anchor cannot have a concrete parent type
EquivalenceClass::getAnchor() always returns a canonical type and
the parent of a canonical type is itself be canonical.

This means that EquivalenceClass::getTypeInContext() can safely
assume that the parent type of a DependentMemberType maps to an
ArchetypeType in the GenericEnvironment.

Instead of trying to handle the case where the parent is concrete,
let's just crash by changing the conditional cast to an
unconditional one.