[test] Invert stdlib_dir conditional sense.
In #32903, a substitution pseudovariable was introduced to refer to the
path containing Swift.swiftmodule. Since builds on Linux put this file
in a path with a subdirectory named after the architecture name -- and
presumably builds on mac OS do not, a conditional was required to
distinguish these two cases.

However, builds on OpenBSD for example also put Swift.swiftmodule in a
similar subdirectory like Linux. Thus, invert the sense of this
conditional to encompass both Linux, OpenBSD, and make mac OS the
specific case, which it is here.