[sil] Add an implementation of isIndirectResultOperand() onto ApplySite that returns false for partial_apply.
Sometimes when working with the operands of an ApplySite, one needs to bail
early if one has a full apply site and the operand is an indirect result
argument. Sadly, there isn't any such API on ApplySite (it is only on
FullApplySite), even though technically a partial_apply can be viewed as just
not having any indirect result operands! That is what this patch implements.

To implement this I added a simple useful method called:

```
Optional<FullApplySite> ApplySite::asFullApplySite() const
```

Seems like a useful thing to have for these sorts of cases.